refactor: lookup, rm defaults (#69)
prep work for #38
