+++ src/store/model/amazon.ts
