+++ src/store/model/evga.ts
