+++ src/store/model/newegg.ts
