+++ src/store/model/nvidia.ts
